feat: done modules modele
